requirements.txt
TgCrypto
ffmpeg-python
git+https://github.com/pyrogram/pyrogram@master
py-tgcalls==0.8.0
python-dotenv
